+++ packages/bluesky/test/src/feeds/feeds_service_test.dart
@@ -1265,4 +1265,74 @@ void main() {
       );
     });
   });
+
+  group('.findActorFeeds', () {
+    test('normal case', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/feeds/data/find_actor_feeds.json',
+        ),
+      );
+
+      final response = await feeds.findActorFeeds(
+        actor: 'shinyakato.dev',
+        limit: 10,
+        cursor: 'xxxxxx',
+      );
+
+      expect(response, isA<XRPCResponse>());
+      expect(response.data, isA<ActorFeeds>());
+    });
+
+    test('when unauthorized', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await feeds.findActorFeeds(
+          actor: 'shinyakato.dev',
+        ),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final feeds = FeedsService(
+        atproto: ATProto(did: 'test', accessJwt: 'test'),
+        protocol: Protocol.https,
+        service: 'test',
+        context: ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedGetClient: atp_test.createMockedGetClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await feeds.findActorFeeds(
+          actor: 'shinyakato.dev',
+        ),
+      );
+    });
+  });
 }